Kill DBAL 2 support
