+++ Tests/Messenger/DoctrinePingConnectionMiddlewareTest.php
@@ -102,13 +102,6 @@ class DoctrinePingConnectionMiddlewareTest extends MiddlewareTestCase
 
     public function testMiddlewareNoPingInNonWorkerContext()
     {
-        // This method has been removed in DBAL 3.0
-        if (method_exists(Connection::class, 'ping')) {
-            $this->connection->expects($this->never())
-                ->method('ping')
-                ->willReturn(false);
-        }
-
         $this->connection->expects($this->never())
             ->method('close')
         ;
