+++ Types/AbstractUidType.php
@@ -88,11 +88,6 @@ abstract class AbstractUidType extends Type
 
     private function hasNativeGuidType(AbstractPlatform $platform): bool
     {
-        // Compatibility with DBAL < 3.4
-        $method = method_exists($platform, 'getStringTypeDeclarationSQL')
-            ? 'getStringTypeDeclarationSQL'
-            : 'getVarcharTypeDeclarationSQL';
-
-        return $platform->getGuidTypeDeclarationSQL([]) !== $platform->$method(['fixed' => true, 'length' => 36]);
+        return $platform->getGuidTypeDeclarationSQL([]) !== $platform->getStringTypeDeclarationSQL(['fixed' => true, 'length' => 36]);
     }
 }
