+++ composer.json
@@ -44,13 +44,13 @@
         "doctrine/annotations": "^1.13.1|^2",
         "doctrine/collections": "^1.0|^2.0",
         "doctrine/data-fixtures": "^1.1",
-        "doctrine/dbal": "^2.13.1|^3.0",
+        "doctrine/dbal": "^3.6",
         "doctrine/orm": "^2.15",
         "psr/log": "^1|^2|^3"
     },
     "conflict": {
         "doctrine/annotations": "<1.13.1",
-        "doctrine/dbal": "<2.13.1",
+        "doctrine/dbal": "<3.6",
         "doctrine/lexer": "<1.1",
         "doctrine/orm": "<2.15",
         "symfony/cache": "<6.4",
